requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
APScheduler==3.10.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.1.0
boto3==1.26.130
botocore==1.29.130
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
decouple==0.0.7
Deprecated==1.2.13
Django==4.2.7
django-cors-headers==3.13.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==22.1
django-triggers==2.1.5
djangorestframework==3.14.0
dnspython==2.5.0
email-validator==2.1.0.post1
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gunicorn==20.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.0
Markdown==3.4.1
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
pdfkit==1.0.0
priority==1.3.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
pyOpenSSL==23.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
redis==4.5.4
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
twilio==8.3.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15
vine==5.0.0
vonage==3.5.1
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.9.2
zipp==3.12.1
zope.interface==6.0
